Rename test_main -> main ; and main -> main_old
